requirements/pypi.txt
discord.py
arrow
asynctk